+++ mrentity/zeronull_int64.go
@@ -10,15 +10,6 @@ type (
 	ZeronullInt64 int64
 )
 
-// Value implements the driver Valuer interface.
-func (n ZeronullInt64) Value() (driver.Value, error) {
-	if n == 0 {
-		return nil, nil
-	}
-
-	return int64(n), nil
-}
-
 // Scan implements the Scanner interface.
 func (n *ZeronullInt64) Scan(value any) error {
 	if value == nil {
@@ -37,4 +28,13 @@ func (n *ZeronullInt64) Scan(value any) error {
 	}
 
 	return mrcore.FactoryErrInternalTypeAssertion.New("ZeronullInt64", value)
+}
+
+// Value implements the driver Valuer interface.
+func (n ZeronullInt64) Value() (driver.Value, error) {
+	if n == 0 {
+		return nil, nil
+	}
+
+	return int64(n), nil
 }