+++ mrfilestorage/errors.go
@@ -1,0 +1,10 @@
+package mrfilestorage
+
+import (
+	"github.com/mondegor/go-sysmess/mrerr"
+)
+
+var (
+	FactoryErrInvalidPath = mrerr.NewFactory(
+		"errMrFileStorageInvalidPath", mrerr.ErrorKindInternal, "invalid path '{{ .path }}'")
+)