+++ mrminio/wrapper_errors.go
@@ -23,12 +23,12 @@ func (fp *FileProvider) wrapError(err error) error {
 	return mrcore.FactoryErrInternal.Caller(1).Wrap(err)
 }
 
-func (fp *FileProvider) debugCmd(ctx context.Context, command, fileName string) {
+func (fp *FileProvider) debugCmd(ctx context.Context, command, filePath string) {
 	mrctx.Logger(ctx).Debug(
 		"%s: cmd=%s, bucket=%s, file=%s",
 		providerName,
 		command,
 		fp.bucketName,
-		fp.baseDir+fileName,
+		filePath,
 	)
 }