+++ mrredsync/locker.go
@@ -54,6 +54,8 @@ func (l *lockerAdapter) LockWithExpiry(ctx context.Context, key string, expiry t
 	}
 
 	return func() {
+		l.debugCmd(ctx, "Unlock", key)
+
 		if _, err := mutex.UnlockContext(ctx); err != nil {
 			mrctx.Logger(ctx).Error(
 				"%s: cmd=unlock, key=%s, err={%s}",
@@ -62,7 +64,5 @@ func (l *lockerAdapter) LockWithExpiry(ctx context.Context, key string, expiry t
 				err,
 			)
 		}
-
-		l.debugCmd(ctx, "Unlock", key)
 	}, nil
 }