+++ mrstorage/file_provider.go
@@ -8,16 +8,9 @@ import (
 
 type (
 	FileProviderAPI interface {
-		Info(ctx context.Context, fileName string) (mrtype.FileInfo, error)
-		Download(ctx context.Context, fileName string) (*mrtype.File, error)
-		// Downloads(ctx context.Context, fileName string) (*mrtype.DownloadedFile, error) // ListObjects :TODO: получение списка объектов
-		Upload(ctx context.Context, file *mrtype.File) error
-		Remove(ctx context.Context, fileName string) error
-	}
-
-	// ExtFileProviderAPI - WARNING: use only when starting the main process
-	ExtFileProviderAPI interface {
-		WithBaseDir(value string) (ExtFileProviderAPI, error)
-		FileProviderAPI
+		Info(ctx context.Context, filePath string) (mrtype.FileInfo, error)
+		Download(ctx context.Context, filePath string) (mrtype.File, error)
+		Upload(ctx context.Context, file mrtype.File) error
+		Remove(ctx context.Context, filePath string) error
 	}
 )